requirements.txt
streamlit
google-cloud-vision
fpdf
Pillow
